refactor: common module에 business 패키지 생성 및 service interface 생성
