+++ common/src/main/java/com/food/common/user/business/AccountFindService.java
@@ -1,0 +1,7 @@
+package com.food.common.user.business;
+
+import javax.validation.constraints.NotNull;
+
+public interface AccountFindService {
+    AccountFindResponse findAccountByUserId(@NotNull Long userId);
+}
