+++ common/src/main/java/com/food/common/user/business/account/impl/DefaultAccountDomainService.java
@@ -1,14 +1,15 @@
-package com.food.common.user.service.account;
+package com.food.common.user.business.account.impl;
 
+import com.food.common.user.business.account.AccountDomainService;
+import com.food.common.user.business.account.impl.response.FoundAppAccount;
+import com.food.common.user.business.account.impl.response.FoundSocialAccount;
+import com.food.common.user.business.user.UserDomainService;
+import com.food.common.user.business.user.impl.response.FoundUser;
 import com.food.common.user.domain.AppAccount;
 import com.food.common.user.domain.SocialAccount;
 import com.food.common.user.domain.User;
 import com.food.common.user.repository.AppAccountRepository;
 import com.food.common.user.repository.SocialAccountRepository;
-import com.food.common.user.service.account.response.FoundAppAccount;
-import com.food.common.user.service.account.response.FoundSocialAccount;
-import com.food.common.user.service.user.UserDomainService;
-import com.food.common.user.service.user.response.FoundUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,10 +19,9 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Transactional
 @Service
-public class AccountDomainService {
+public class DefaultAccountDomainService implements AccountDomainService {
     private final AppAccountRepository appAccountRepository;
     private final SocialAccountRepository socialAccountRepository;
-
     private final UserDomainService userDomainService;
 
     public Optional<FoundAppAccount> findAppAccountByUserId(Long userId) {
