+++ common/src/main/java/com/food/common/user/business/user/UserDomainService.java
@@ -1,0 +1,9 @@
+package com.food.common.user.business.user;
+
+import com.food.common.user.business.user.impl.response.FoundUser;
+import com.food.common.user.domain.User;
+
+public interface UserDomainService {
+    FoundUser getFoundUser(User user);
+    User findEntityById(Long id);
+}
